First renderable version!
1. Added a filter Geometry to group up geometry releated classes.
2. Added information describing the distance between the intersection
point and the camera to class Intersection.
3. Modified method Intersect in interface IIntersectTarget and all its
sub-classes since we only need the cloest intersection instead of all of
them.
4. Added two method to deal with object color to interface
IIntersectTarget and a brand new base class which inherits from it
implementing these methods became super class for all the sub-classes of
it.
5. Fixed the color issue for saving an Image object as ppm file.
6. Added the render classes. One is a abstract base class while the
other is an implementor implementing it with ray tracing.
7. Added class Scene to contain a set of renderable objects and a set of
cameras.