Octree!
1. An interface IScene which inherits from IIntersectTarget making its
sub-classes able to do intersect testing for all kinds of scenes.
2. Moved Raytracer's intersection test code to SimpleScene which now do
the same as previously what the RayTracer did.
3. Added a new kind of scene, OctreeScene, using octree to manage the
objects within a scene. This improves the performance quite a bit.
4. Added Octree\OctreeNode to package Utils as the foundation of
OctreeScene.
5. Now all the bounding volume (e.q. AABB and Volume) uses a new
interface IBoundingVolume which enables them to be extended or return
the minimal extent and the maximal extent as their parent class instead
of Primitive which obviously is no use for them.
6. Moved the definition for List to Defs.h making it available to more
files.
7. Added a method to return an Object's bounding volume.
8. Now scenes require to be initialize before being used.
9. Fixed two little bugs for comparing two Vector objects.